diffusion.py
import torch
from diffusers import StableDiffusionInpaintPipeline
from PIL import Image
import requests
import numpy as np

# Load a pretrained stable diffusion model from Hugging Face
model_id = "CompVis/stable-diffusion-v1-4"
pipe = StableDiffusionInpaintPipeline.from_pretrained(model_id)
pipe = pipe.to("cuda")  # Use GPU for faster inference

# Load an example damaged image
url = "https://example.com/damaged_image.jpg"
response = requests.get(url)
damaged_image = Image.open(BytesIO(response.content))

# Perform inpainting to restore the image
prompt = "A restored version of the damaged image"
restored_image = pipe(prompt, image=damaged_image).images[0]

# Save the restored image
restored_image.save("restored_image.jpg")