requirements.txt
torch
diffusers
opencv-python
numpy
matplotlib
pillow
requests